Started to implement damage dealing

+++ DeepSpace.Contracts/Ship.cs
@@ -17,7 +17,7 @@ namespace DeepSpace.Contracts
         public string CommandCode { get; set; }
         public string TransponderCode { get; set; }
         
-        public double Shield => ShieldUpgrades.Sum(su => su.ShieldValue);
+        public int Shield => ShieldUpgrades.Sum(su => su.ShieldValue);
         public List<IShieldUpgrades> ShieldUpgrades { get; set; }        
 
         public Location Location { get; set; }
